fix(metrics): align success-status policy with documented behavior
